// src/types/profile.ts
/**
 * Interface for user profile data
 */
export interface ProfileData {
  username: string
  full_name: string | null
  avatar_url: string | null
  created_at: string
  github_username: string | null
  twitter_username: string | null
  instagram_username: string | null
  youtube_username: string | null
} 